reorganizes files into new folders
