+++ stories/molecules/buttons/DropdownButton.stories.tsx
@@ -1,12 +1,12 @@
 import { createMeta } from "stories/utils"
-import DropdownButton from "../../../../components/DropdownButton/DropdownButton"
+import DropdownButton from "components/DropdownButton/DropdownButton"
 import { ComponentStory } from "@storybook/react"
 
 // TODO: move into components directory
 //const DropdownButton = () => <div>TODO</div>
 
 export default createMeta({
-  title: "Components/Buttons/DropdownButton",
+  title: "Molecules/Buttons/DropdownButton",
   figmaUrl:
     "https://www.figma.com/file/3ifz37EOwDfmnEG8320KlD/CS1---MAPLE?node-id=245%3A12483",
   component: DropdownButton
