+++ stories/molecules/buttons/LoadingButton.stories.tsx
@@ -10,10 +10,10 @@ import {
 import { Image, Stack } from "components/bootstrap"
 import { useState } from "react"
 import { createMeta } from "stories/utils"
-import { LoadingButton } from "../../../../components/buttons"
+import { LoadingButton } from "components/buttons"
 
 export default createMeta({
-  title: "Components/Buttons/LoadingButton",
+  title: "Molecules/Buttons/LoadingButton",
   component: LoadingButton,
   parameters: {
     docs: {
