+++ stories/molecules/buttons/PrimaryButtonWithIcon.stories.tsx
@@ -20,7 +20,7 @@ const pageArgs: PrimaryButtonWithIconProps = {
 }
 
 export default createMeta({
-  title: "Components/Buttons/PrimaryButtonWithIcon",
+  title: "Molecules/Buttons/PrimaryButtonWithIcon",
   figmaUrl:
     "https://www.figma.com/file/3ifz37EOwDfmnEG8320KlD/CS1---MAPLE?node-id=245%3A14373",
   component: PrimaryButtonWithIcon,
