+++ stories/molecules/readme.md
@@ -1,0 +1,23 @@
+# organisms folder
+
+> In interfaces, molecules are relatively simple groups of UI elements functioning together as a unit. For example, a form label, search input, and button can join together to create a search form molecule. <br/>
+> When combined, these abstract atoms suddenly have purpose. The label atom now defines the input atom. Clicking the button atom now submits the form. The result is a simple, portable, reusable component that can be dropped in anywhere search functionality is needed.
+> <br/>[...]<br/>
+> [C]reating simple UI molecules makes testing easier, encourages reusability, and promotes consistency throughout the interface.
+> <br />
+> Now we have simple, functional, reusable components that we can put into a broader context.
+
+<div style="text-align: right">
+brad frost <br />
+atomic design <br />
+https://atomicdesign.bradfrost.com/chapter-2/#molecules
+</div>
+
+### what goes here?
+
+- Simple groups of atom components that are more than the sum of their parts.
+- Simple groups of atom components that as a unit performs one function. 
+- Simple groups of atom components that classically 'belong' together, appear to the user as a unit, are treated in our design as a unit, and probably are reused very frequently, like a label and a form input. 
+
+
+If this folder gets too big, we can sort these into subfolders by type and function.
