+++ stories/organisms/cards/CardListItem.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory } from "@storybook/react"
 import { createMeta } from "stories/utils"
-import { CardListItems, ListItem } from "../../../../components/Card"
+import { CardListItems, ListItem } from "components/Card"
 
 export default createMeta({
   title: "Components/Cards/CardListItem",
@@ -17,6 +17,7 @@ export const Primary = Template.bind({})
 Primary.args = {
   items: [
     <ListItem
+      key={1}
       billName="H.3330"
       billDescription="Important bill, must vote!"
       element={undefined}
@@ -26,14 +27,14 @@ Primary.args = {
 
 export const OnlyBillName = Template.bind({})
 OnlyBillName.args = {
-  items: [<ListItem billName="H.3330" />]
+  items: [<ListItem key={1} billName="H.3330" />]
 }
 
 export const MultipleItems = Template.bind({})
 MultipleItems.args = {
   items: [
-    <ListItem billName="H.3330" billDescription="Important bill, must vote!" />,
-    <ListItem billName="H.3330" billDescription="Important bill, must vote!" />,
-    <ListItem billName="H.3330" billDescription="Important bill, must vote!" />
+    <ListItem key={1} billName="H.3330" billDescription="Important bill, must vote!" />,
+    <ListItem key={2} billName="H.3330" billDescription="Important bill, must vote!" />,
+    <ListItem key={3} billName="H.3330" billDescription="Important bill, must vote!" />
   ]
 }
