+++ stories/organisms/newsfeed/AlertCard.stories.tsx
@@ -5,7 +5,7 @@ import { AlertCard } from "components/AlertCard/AlertCard"
 import { Timestamp } from "firebase/firestore"
 
 export default createMeta({
-  title: "Dashboard/Newsfeed/AlertCard",
+  title: "Newsfeed/AlertCard",
   figmaUrl:
     "https://www.figma.com/file/3ifz37EOwDfmnEG8320KlD/CS1---MAPLE?node-id=158%3A2873",
   component: AlertCard
