+++ stories/organisms/readme.md
@@ -1,7 +1,5 @@
 # organisms folder
 
-
-
 > Organisms are relatively complex UI components composed of groups of molecules and/or atoms and/or other organisms. These organisms form distinct sections of an interface.
 
 <div style="text-align: right">
