+++ stories/pages/auth/OrgSignUpModal.stories.tsx
@@ -1,30 +1,30 @@
 import { ComponentStory } from "@storybook/react"
 import { createMeta } from "stories/utils"
-import ProfileTypeModal from "../../../../components/auth/ProfileTypeModal"
+import OrgSignUpModal from "components/auth/OrgSignUpModal"
 import { Providers } from "components/providers"
 import { wrapper } from "components/store"
 import { Provider as Redux } from "react-redux"
 
 export default createMeta({
-  title: "Auth/Profile Type Modal",
+  title: "Auth/Org Sign Up Modal",
   figmaUrl:
     "https://www.figma.com/file/3ifz37EOwDfmnEG8320KlD/CS1---MAPLE?node-id=49%3A3010",
-  component: ProfileTypeModal
+  component: OrgSignUpModal
 })
 
-const Template: ComponentStory<typeof ProfileTypeModal> = args => (
-  <ProfileTypeModal {...args} />
+const Template: ComponentStory<typeof OrgSignUpModal> = args => (
+  <OrgSignUpModal {...args} />
 )
 
 export const Primary = Template.bind({})
 
 Primary.args = {
-  onIndividualUserClick: () => {},
-  onOrgUserClick: () => {},
+  onSuccessfulSubmit: () => {},
   onHide: () => {},
   show: true
 }
-Primary.storyName = "Profile Type Modal"
+Primary.storyName = "Org Sign Up Modal"
+
 Primary.decorators = [
   (Story, ...rest) => {
     const { store, props } = wrapper.useWrappedStore(...rest)
